backend/requirements.txt
flask==3.0.3
flask-cors==4.0.0
python-docx==0.8.11
openai==1.30.0
httpx==0.27.2
zhipuai==2.1.0
requests==2.31.0
python-dotenv==1.0.0
werkzeug==3.0.2
reportlab==4.0.4